[fix] Fixed detection of areas
The area detection now works. Whenever robot is in a defined boost area
it gets detected.
slam_pose is now only used indirectly to update the state.
It could be changed in the way that /tf gets subscribed to instead.
This is not necessary though, because it worked in exactly the same
way, only that it were more intuitive.

Next up: Define a new topic and