[add] Added visualization of boost rectangles
This commit is quite large. It tries to add visualization of the rectangles that define the
boost areas. The Rectangles shall be drawn in rviz by a topic which includes 6 Poses per rectangle.

Rviz just needs to be configured so that the PoseArray get drawn.

Problem now:	Arrows are not as they are supposed to be. The problem is assumed to be the math.
		It shall be reviewed and checked for mistakes soon.