[add] Added node for recording and replaying waypoints
The new node shall be able to record a waypoint whenever the user presses A-Button during the manual
driving round in the presentation. In the second round, when the car shall drive autonomously,
it successively goes to the respective waypoints until the round is complete. The ideo is to get a nice behaviour in the curves